feat(EPMGCIPUVA-2): Update app structure
