+++ packages/frontend/src/App.module.sass
@@ -1,4 +1,4 @@
-@use './colors'
+@use 'src/style/colors'
 
 .app 
   padding-top: 46px
