+++ packages/frontend/src/components/Button/Button.module.sass
@@ -1,4 +1,4 @@
-@use 'src/colors'
+@use 'src/style/colors'
 
 .button 
   background-color: inherit
