+++ packages/frontend/src/components/Header/Header.module.sass
@@ -1,5 +1,5 @@
-@use 'src/colors'
-@import 'src/typography'
+@use 'src/style/colors'
+@import 'src/style/typography'
 
 .headerWrapper 
     width: 100%
