+++ packages/frontend/src/components/Input/Input.module.sass
@@ -1,5 +1,5 @@
-@use 'src/colors'
-@import 'src/typography'
+@use 'src/style/colors'
+@import 'src/style/typography'
 
 .label
     @extend .H6
