+++ packages/frontend/src/components/Loader/Loader.module.sass
@@ -1,4 +1,4 @@
-@use 'src/colors'
+@use 'src/style/colors'
 
 .loader
     border: 6px solid colors.$blue-dark
