+++ packages/frontend/src/components/Menu/Menu.module.sass
@@ -1,5 +1,5 @@
-@use 'src/colors'
-@import 'src/typography'
+@use 'src/style/colors'
+@import 'src/style/typography'
 
 .menuWrapper
     z-index: -1
