+++ packages/frontend/src/components/Notification/Notification.module.sass
@@ -1,5 +1,5 @@
-@use 'src/colors'
-@import 'src/typography'
+@use 'src/style/colors'
+@import 'src/style/typography'
 
 .notificationWrapper
     @extend .H6
