+++ packages/frontend/src/pages/CreateUserPage/CreateUserPage.module.sass
@@ -1,5 +1,5 @@
-@use 'src/colors'
-@import 'src/typography'
+@use 'src/style/colors'
+@import 'src/style/typography'
 
 .createUserWrapper
     padding: 24px
