+++ packages/frontend/src/pages/ManageUsersPage/ManageUsersPage.module.sass
@@ -1,4 +1,4 @@
-@use 'src/colors'
+@use 'src/style/colors'
 
 .manageUsersWrapper
     padding: 16px
