+++ packages/frontend/src/pages/TitlePage/TitlePage.module.sass
@@ -1,4 +1,4 @@
-@use 'src/colors'
+@use 'src/style/colors'
 
 .titleWrapper 
   width: 100%
